# Enhance Solving Problems

+++ src/_3_problems_from_21_to_30/_3_10_problem_30/Problem30.cpp
@@ -18,11 +18,11 @@ char readCharacter() {
     return character;
 }
 
-unsigned long long targetLetterCount(
+size_t targetLetterCount(
     const string& TEXT,
     const char& TARGET_CHARACTER
 ) {
-    unsigned long long counter = 0;
+    size_t counter = 0;
     for (const char& CHARACTER : TEXT)
         if (CHARACTER == TARGET_CHARACTER)
             counter++;
